Added Additional  Brightness Slider Information for System Presets
Added additional system preset brightness slider information for downloadable presets. Also corrected a minor typo with "Tooltip width" to "Tooltip Width". Feature update Compiled without errors/ran, but because the app is looking for the json file from the web it doesn't load this new information until that file gets updated on the repo. Looking into what needs to be done to load a local file as a uri for future testing.