Added Additional Brightness Slider Information for System Presets (#26)
* Added Image Profile support for 4K HDR small font English

This is a basic image profile for 4K (3840x2160) HDR with small game font in English. It has most item types referenced (missing a few), and affixes/aspects needed for Twisting Blades Rogue. More reference images need to be captured to support other classes/builds.

* Added Additional  Brightness Slider Information for System Presets

Added additional system preset brightness slider information for downloadable presets. Also corrected a minor typo with "Tooltip width" to "Tooltip Width". Feature update Compiled without errors/ran, but because the app is looking for the json file from the web it doesn't load this new information until that file gets updated on the repo. Looking into what needs to be done to load a local file as a uri for future testing.

* Update systempresets.json

added line 31 back that got accidentally deleted.

* Update systempresets.json

Adjusted the BrightnessSlider values to be easier to read when displayed. An example:  "60 255" => "60 / 255".

* Update SettingsView.xaml

Fixed a typo on line 106 referencing the wrong grid on the ui.
Changed from 3 to 5.

* Some naming and gui changes
---------

Co-authored-by: Jos Demmers <josdemmers@gmail.com>